+++ starlette/middleware/base.py
@@ -1,4 +1,5 @@
 import typing
+import warnings
 
 import anyio
 
@@ -12,6 +13,14 @@ DispatchFunction = typing.Callable[
     [Request, RequestResponseEndpoint], typing.Awaitable[Response]
 ]
 T = typing.TypeVar("T")
+
+warnings.warn(
+    "The 'BaseHTTPMiddleware' is deprecated, and will be removed in version 2.0.0."
+    "Refer to https://www.starlette.io/middleware/#pure-asgi-middleware to learn "
+    "how to create middlewares.\nIf you need help, please create a discussion on: "
+    "https://github.com/encode/starlette/discussions.",
+    DeprecationWarning,
+)
 
 
 class _CachedRequest(Request):